# api/openapi.yaml
openapi: 3.1.0
info:
  title: 책결 API
  version: 0.7.0
  description: 개인 로컬 계정, 친구·그룹 독서 공유, 기록, 통계, 알림, 계정 삭제와 운영자 안전 기능을 제공하는 현재 Go 서버 계약.
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
  - devUserAuth: []
...
paths:
...
  /v1/auth/register:
    post:
      tags: [Authentication]
      summary: 최초 개인 계정 생성
      description: local_credentials가 비어 있을 때 한 번만 성공한다.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LocalRegistration' }
      responses:
        '201':
          description: 생성된 계정과 30일 세션
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LocalAuthSession' }
        '400': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
  /v1/auth/login:
    post:
      tags: [Authentication]
      summary: 개인 계정 로그인
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LocalLogin' }
      responses:
        '200':
          description: 새 30일 세션
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LocalAuthSession' }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
  /v1/auth/logout:
    post:
      tags: [Authentication]
      summary: 현재 개인 세션 폐기
      responses:
        '204': { $ref: '#/components/responses/NoContent' }
        '401': { $ref: '#/components/responses/Error' }
  /v1/me/bootstrap:
    post:
      tags: [Account]
      summary: 인증 사용자의 서비스 프로필 보장
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                nickname: { type: string, maxLength: 40 }
      responses:
        '204': { $ref: '#/components/responses/NoContent' }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
  /v1/me:
    get:
      tags: [Account]
      summary: 내 프로필 조회
      responses:
        '200':
          description: 내 프로필
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Profile' }
        '401': { $ref: '#/components/responses/Error' }
    patch:
      tags: [Account]
      summary: 내 프로필과 기본 공개 설정 수정
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateProfile' }
      responses:
        '200':
          description: 수정된 프로필
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Profile' }
        '400': { $ref: '#/components/responses/Error' }
    delete:
      tags: [Account]
      summary: 로그인 계정과 서비스 데이터 삭제
      description: 개인 계정은 PostgreSQL 자격 증명과 세션까지 함께 삭제한다. Supabase 연동 모드에서는 외부 Auth 삭제 실패를 Worker가 재시도한다.
      responses:
        '204': { $ref: '#/components/responses/NoContent' }
        '202':
          description: 외부 Auth 또는 로컬 삭제 재시도 대기
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status: { const: pending }
        '401': { $ref: '#/components/responses/Error' }
  /v1/me/storage-status:
    get:
      tags: [Account]
      summary: 내 PostgreSQL 저장 상태와 기록 건수 확인
      responses:
        '200':
          description: 저장 상태
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StorageStatus' }
        '401': { $ref: '#/components/responses/Error' }
  /v1/me/stats:
    get:
      tags: [Account]
      summary: 연도별 독서 통계 조회
      parameters:
        - name: year
          in: query
          schema: { type: integer, minimum: 2000, maximum: 2200 }
      responses:
        '200':
          description: 독서 통계
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ReadingStats' }
        '400': { $ref: '#/components/responses/Error' }
  /v1/me/notifications:
    get:
      tags: [Account]
      summary: 알림 설정 조회
      responses:
        '200':
          description: 알림 설정
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NotificationPreferences' }
    put:
      tags: [Account]
      summary: 알림 설정 전체 교체
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/NotificationPreferences' }
      responses:
        '200':
          description: 저장된 알림 설정
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NotificationPreferences' }
        '400': { $ref: '#/components/responses/Error' }
  /v1/me/push-tokens:
    post:
      tags: [Account]
      summary: 기기 푸시 토큰 등록 또는 갱신
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PushToken' }
      responses:
        '204': { $ref: '#/components/responses/NoContent' }
        '400': { $ref: '#/components/responses/Error' }
    delete:
      tags: [Account]
      summary: 내 모든 기기 푸시 토큰 비활성화
      responses:
        '204': { $ref: '#/components/responses/NoContent' }
  /v1/admin/reports:
    get:
      tags: [Moderation]
      summary: 신고 큐 조회
      security: [{ adminKeyAuth: [] }]
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [all, open, reviewing, resolved, dismissed]
            default: open
        - $ref: '#/components/parameters/AdminLimit'
        - $ref: '#/components/parameters/AdminIDHeader'
      responses:
        '200':
          description: 신고 목록
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/AdminReport' }
        '401': { $ref: '#/components/responses/Error' }
  /v1/admin/reports/{reportId}:
    patch:
      tags: [Moderation]
      summary: 신고 기각·숨김·경고·정지·차단 처리
      security: [{ adminKeyAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/ReportID'
        - $ref: '#/components/parameters/AdminIDHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ResolveReport' }
      responses:
        '200':
          description: 생성된 감사 작업
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ModerationAction' }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
        '422': { $ref: '#/components/responses/Error' }
  /v1/admin/hidden-targets/{targetType}/{targetId}/restore:
    post:
      tags: [Moderation]
      summary: 운영자 숨김 대상 복원
      security: [{ adminKeyAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/ModerationTargetType'
        - $ref: '#/components/parameters/ModerationTargetID'
        - $ref: '#/components/parameters/AdminIDHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason: { type: string, minLength: 1, maxLength: 1000 }
      responses:
        '200':
          description: 생성된 감사 작업
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ModerationAction' }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
  /v1/admin/moderation/actions:
    get:
      tags: [Moderation]
      summary: 최근 운영 조치 감사 로그 조회
      security: [{ adminKeyAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/AdminLimit'
        - $ref: '#/components/parameters/AdminIDHeader'
      responses:
        '200':
          description: 감사 로그
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/ModerationAction' }
        '401': { $ref: '#/components/responses/Error' }
  /v1/me/annual-goal:
    put:
      tags: [Account]
      summary: 연간 완독 목표 설정
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [year, targetBooks]
              properties:
                year: { type: integer, minimum: 2000, maximum: 2200 }
                targetBooks: { type: integer, minimum: 1, maximum: 1000 }
      responses:
        '204': { $ref: '#/components/responses/NoContent' }
        '400': { $ref: '#/components/responses/Error' }
...
  /v1/me/reading-presence:
    put:
      tags: [Account]
      summary: 현재 독서 중 상태 시작
      description: 상태는 서버에서 최대 2시간 뒤 만료된다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [readingRunId]
              properties:
                readingRunId: { type: string, format: uuid }
      responses:
        '204': { $ref: '#/components/responses/NoContent' }
        '404': { $ref: '#/components/responses/Error' }
    delete:
      tags: [Account]
      summary: 현재 독서 중 상태 종료
      responses:
        '204': { $ref: '#/components/responses/NoContent' }
...
  /v1/catalog/books:
    get:
      tags: [Catalog]
      summary: 제목·저자·ISBN 도서 검색
      parameters:
        - name: query
          in: query
          required: true
          schema: { type: string, minLength: 2, maxLength: 100 }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 50, default: 20 }
      responses:
        '200':
          description: 검색 결과
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Book' }
        '400': { $ref: '#/components/responses/Error' }
        '503': { $ref: '#/components/responses/Error' }
  /v1/catalog/books/{isbn}:
    get:
      tags: [Catalog]
      summary: ISBN으로 도서 조회
      parameters:
        - $ref: '#/components/parameters/ISBN'
      responses:
        '200':
          description: 도서
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Book' }
        '400': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
  /v1/reading-runs:
    get:
      tags: [Reading]
      summary: 내 모든 독서 회차 조회
      responses:
        '200':
          description: 독서 회차
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/ReadingRun' }
    post:
      tags: [Reading]
      summary: 카탈로그 도서로 독서 회차 생성
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateReadingRun' }
      responses:
        '201':
          description: 생성된 독서 회차
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ReadingRun' }
        '400': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
  /v1/reading-runs/manual:
    post:
      tags: [Reading]
      summary: 도서를 직접 입력해 독서 회차 생성
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateManualReadingRun' }
      responses:
        '201':
          description: 생성된 독서 회차
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ReadingRun' }
        '400': { $ref: '#/components/responses/Error' }
  /v1/reading-runs/{readingRunId}:
    patch:
      tags: [Reading]
      summary: 독서 상태와 공유 정책 수정
      parameters:
        - $ref: '#/components/parameters/ReadingRunID'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateReadingRun' }
      responses:
        '200':
          description: 수정된 독서 회차
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ReadingRun' }
        '400': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
  /v1/reading-runs/{readingRunId}/entries:
    get:
      tags: [Reading]
      summary: 독서 회차의 진척 기록 조회
      parameters:
        - $ref: '#/components/parameters/ReadingRunID'
      responses:
        '200':
          description: 진척 기록
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/ProgressEntry' }
        '404': { $ref: '#/components/responses/Error' }
  /v1/reading-runs/{readingRunId}/progress:
    post:
      tags: [Reading]
      summary: 진척·메모·독서 시간 기록
      parameters:
        - $ref: '#/components/parameters/ReadingRunID'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RecordProgress' }
      responses:
        '201':
          description: 새 진척 기록
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RecordProgressResult' }
        '200':
          description: 같은 clientOperationId의 멱등 재생
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RecordProgressResult' }
        '409': { $ref: '#/components/responses/Error' }
        '422': { $ref: '#/components/responses/Error' }
  /v1/feed:
    get:
      tags: [Social]
      summary: 관계·차단·공개 범위가 적용된 피드 조회
      parameters:
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 50, default: 30 }
      responses:
        '200':
          description: 피드
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/FeedEvent' }
  /v1/feed-events/{feedEventId}/reaction:
    parameters:
      - $ref: '#/components/parameters/FeedEventID'
    put:
      tags: [Social]
      summary: 피드에 응원 추가
      responses:
        '204': { $ref: '#/components/responses/NoContent' }
        '404': { $ref: '#/components/responses/Error' }
    delete:
      tags: [Social]
      summary: 피드 응원 취소
      responses:
        '204': { $ref: '#/components/responses/NoContent' }
        '404': { $ref: '#/components/responses/Error' }
...
  /v1/friend-invites/{token}/accept:
    post:
      tags: [Social]
      summary: 친구 초대 수락
      parameters:
        - $ref: '#/components/parameters/InviteToken'
      responses:
        '200':
          description: 연결된 친구
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Friend' }
        '404': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque local session or Supabase JWT
    devUserAuth:
      type: apiKey
      in: header
      name: X-User-ID
      description: ALLOW_DEV_AUTH=true인 로컬 개발에서만 허용한다.
    adminKeyAuth:
      type: apiKey
      in: header
      name: X-Admin-Key
      description: 서버 비밀 저장소의 ADMIN_API_KEY와 일치해야 한다.
  parameters:
...
    AdminLimit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
...
  responses:
    NoContent:
      description: 처리 완료
    Error:
      description: 오류 응답
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
...
    LocalRegistration:
      type: object
      additionalProperties: false
      required: [email, password, nickname]
      properties:
        email: { type: string, format: email, maxLength: 254 }
        password: { type: string, minLength: 10, maxLength: 72 }
        nickname: { type: string, minLength: 1, maxLength: 40 }
    LocalLogin:
      type: object
      additionalProperties: false
      required: [email, password]
      properties:
        email: { type: string, format: email, maxLength: 254 }
        password: { type: string, minLength: 10, maxLength: 72 }
...
    LocalAuthSession:
      type: object
      required: [token, expiresAt, user]
      properties:
        token: { type: string, writeOnly: true }
        expiresAt: { type: string, format: date-time }
        user: { $ref: '#/components/schemas/LocalAuthUser' }
...